Update 4/17/2023
- Enforcement of API Keys
API Keys have been released for a week, Grab yours now!!! (#📊・api)
We are soon going to be mandating the use of **API Keys**! These keys __will NOT__ replace the use of bot tokens, But will be used to stop malicious users from gaining access to places they shouldn't be in!


- Introducing: BDFDData-Database
We now allow users to have their own Database! This will help with BDFD/Website Dashboard connections, and so much more!
Endpoints are listed below.
- `/db/set` - Set a variable
- `/db/get`  - Get a variable
- `/db/getAll` - Get all variables
- `/db/delete` - delete a variable
- `/db/deleteAll` - delete all variables


- New and Exciting Endpoints 
- `/guild/prune` Prune Guild Member(s)
- `/guild/delete-emoji` Delete Guild Emoji(s)
- `/guild/bans` - Get Guild Ban(s)
- `/guild/event` - Get Guild Event(s) 
- `/guild/event/create` - Create Guild Event
- `/guild/event/delete` - Delete Guild Event
- `/guild/event/modify` - Modify Guild Event 
- `/guild/event/users` - Get Interested Users for Guild Event 
- `/guild/invite` - Get Guild Invite(s)
- `/reaction` - Get reaction(s) 
- `/reaction/add` - adds reaction 
- `/reaction/delete` - deletes reaction(s)


- Important Changes 
★ The following endpoints have been changed and **no longer support using names, and can now only be identified with ID's** (more info in docs)
- `guild/emojis`
- `guild/channels`
- `guild/roles`
- `guild/members`
- `guild/slash`

★ All endpoints with optional parameters **__no longer strictly send arrays__**.  (more info in docs)


- Removal of Custom Endpoint 
Unfortunately, We have made the decision to remove `/custom` and `/parse/time` from our API. We know how much this will affect users and cause inconveniences. But it was too dangerous and broken.